+++ python/ry/__about__.py
@@ -2,7 +2,13 @@
 
 from __future__ import annotations
 
-from ry.ryo3 import __authors__, __build_profile__, __build_timestamp__, __version__
+from ry.ryo3 import (
+    __authors__,
+    __build_profile__,
+    __build_timestamp__,
+    __target__,
+    __version__,
+)
 
 __all__ = (
     "__authors__",
@@ -10,6 +16,7 @@ __all__ = (
     "__build_timestamp__",
     "__description__",
     "__pkgroot__",
+    "__target__",
     "__title__",
     "__version__",
 )